// src/shared/types/MovieSummary.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type MovieSummary = { tmdbId: bigint, title: string, originalTitle: string, poster: string | null, backdrop: string | null, releaseDate: string | null, overview: string, genreIds: Array<number>, rating: number, voteCount: bigint, popularity: number, };